+++ common/src/main/java/traben/entity_model_features/mixin/MixinModelPart$Cuboid.java
@@ -6,13 +6,13 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import traben.entity_model_features.EMF;
-import traben.entity_model_features.utils.EMFCuboidDataSupplier;
+import traben.entity_model_features.utils.IEMFCuboidDataSupplier;
 
 import java.util.Set;
 import net.minecraft.client.model.geom.ModelPart;
 
 @Mixin(ModelPart.Cube.class)
-public class MixinModelPart$Cuboid implements EMFCuboidDataSupplier {
+public class MixinModelPart$Cuboid implements IEMFCuboidDataSupplier {
 
 
     @Unique