+++ common/src/main/java/traben/entity_model_features/mixin/MixinModelPartData.java
@@ -7,15 +7,15 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import traben.entity_model_features.EMF;
-import traben.entity_model_features.utils.EMFTextureSizeSupplier;
+import traben.entity_model_features.utils.IEMFTextureSizeSupplier;
 
 @Mixin(PartDefinition.class)
 public class MixinModelPartData {
     @Inject(method = "bake",
             at = @At(value = "RETURN"))
     private void emf$injectAnnouncer(final int textureWidth, final int textureHeight, final CallbackInfoReturnable<ModelPart> cir) {
-        if (EMF.config().getConfig().modelExportMode.doesLog() && cir.getReturnValue() instanceof EMFTextureSizeSupplier) {
-            ((EMFTextureSizeSupplier) cir.getReturnValue()).emf$setTextureSize(new int[]{textureWidth, textureHeight});
+        if (EMF.config().getConfig().modelExportMode.doesLog() && cir.getReturnValue() instanceof IEMFTextureSizeSupplier) {
+            ((IEMFTextureSizeSupplier) cir.getReturnValue()).emf$setTextureSize(new int[]{textureWidth, textureHeight});
         }
 
     }