+++ common/src/main/java/traben/entity_model_features/mixin/rendering/arrows/MixinProjectileEntityRenderer.java
@@ -16,7 +16,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import traben.entity_model_features.models.animation.EMFAnimationEntityContext;
-import traben.entity_model_features.utils.EMFCustomModelHolder;
+import traben.entity_model_features.utils.IEMFCustomModelHolder;
 
 @Mixin(ArrowRenderer.class)
 public abstract class MixinProjectileEntityRenderer<T extends AbstractArrow> extends EntityRenderer<T> {
@@ -30,7 +30,7 @@ public abstract class MixinProjectileEntityRenderer<T extends AbstractArrow> ext
     @Inject(method = "render(Lnet/minecraft/world/entity/projectile/AbstractArrow;FFLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource;I)V",
             at = @At(value = "INVOKE", target = "Lcom/mojang/blaze3d/vertex/PoseStack;last()Lcom/mojang/blaze3d/vertex/PoseStack$Pose;", shift = At.Shift.BEFORE), cancellable = true)
     private void emf$cancelAndCEMRender(final T entity, final float entityYaw, final float partialTicks, final PoseStack poseStack, final MultiBufferSource buffer, final int packedLight, final CallbackInfo ci) {
-        if (this instanceof EMFCustomModelHolder customModelHolder && customModelHolder.emf$hasModel()) {
+        if (this instanceof IEMFCustomModelHolder customModelHolder && customModelHolder.emf$hasModel()) {
             //matrixStack.translate(4,0,0);
             poseStack.scale(16, -12.8f, -12.8f);//result 0.9,  0.72   0.72
             EMFAnimationEntityContext.setHeadYaw(entityYaw);