+++ common/src/main/java/traben/entity_model_features/models/EMFModelPartVanilla.java
@@ -83,4 +83,5 @@ public class EMFModelPartVanilla extends EMFModelPartWithState {
     public String toStringShort() {
         return "[vanilla part " + name + "]";
     }
+
 }