+++ common/src/main/java/traben/entity_model_features/models/animation/EMFAnimationEntityContext.java
@@ -3,7 +3,11 @@ package traben.entity_model_features.models.animation;
 import it.unimi.dsi.fastutil.ints.IntArraySet;
 import it.unimi.dsi.fastutil.ints.IntSet;
 import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
+import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
+import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet;
+import it.unimi.dsi.fastutil.objects.ObjectSet;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.core.BlockPos;
@@ -17,12 +21,13 @@ import net.minecraft.world.entity.animal.Parrot;
 import net.minecraft.world.entity.animal.Wolf;
 import net.minecraft.world.entity.animal.camel.Camel;
 import net.minecraft.world.entity.monster.*;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.Projectile;
 import net.minecraft.world.entity.vehicle.Boat;
 import net.minecraft.world.entity.vehicle.Minecart;
 import net.minecraft.world.level.dimension.BuiltinDimensionTypes;
 import net.minecraft.world.phys.Vec3;
-import org.spongepowered.asm.mixin.gen.Accessor;
+import org.jetbrains.annotations.Nullable;
 import traben.entity_model_features.EMF;
 import traben.entity_model_features.mixin.accessor.EntityRenderDispatcherAccessor;
 import traben.entity_model_features.mixin.accessor.MinecraftClientAccessor;
@@ -34,10 +39,11 @@ import traben.entity_texture_features.ETF;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
+import java.util.function.Consumer;
 import java.util.function.Function;
 
 @SuppressWarnings({"resource", "SameParameterValue", "unused"})
-public class EMFAnimationEntityContext {
+public final class EMFAnimationEntityContext {
 
     private static final Object2IntOpenHashMap<UUID> knownHighestAngerTimeByUUID = new Object2IntOpenHashMap<>() {{
         defaultReturnValue(0);
@@ -49,7 +55,8 @@ public class EMFAnimationEntityContext {
     public static boolean setIsOnHead = false;
     public static double lastFOV = 70;
     public static boolean is_in_ground_override = false;
-    private static EMFEntity emfEntity = null;
+    public static ObjectSet<UUID> entitiesToForceVanillaModel = new ObjectOpenHashSet<>();
+    private static EMFEntity IEMFEntity = null;
     private static float shadowSize = Float.NaN;
     private static float shadowOpacity = Float.NaN;
     private static float leashX = Float.NaN;
@@ -66,26 +73,40 @@ public class EMFAnimationEntityContext {
     private static Boolean lodFrameSkipping = null;
     private static boolean announceModels = false;
 
+    public static Object2ObjectOpenHashMap<UUID, ModelPart[]> entitiesPausedParts = new Object2ObjectOpenHashMap<>();
+    public static ObjectSet<UUID> entitiesPaused = new ObjectOpenHashSet<>();
+
+    public static boolean isEntityAnimPaused(){
+        if (IEMFEntity == null) return false;
+        return entitiesPaused.contains(IEMFEntity.etf$getUuid());
+    }
+
+    public static @Nullable ModelPart[] getEntityPartsAnimPaused(){
+        if (IEMFEntity == null) return null;
+        var parts = entitiesPausedParts.get(IEMFEntity.etf$getUuid());
+        return parts.length == 0 ? null : parts;
+    }
+
     private EMFAnimationEntityContext() {
 
     }
 
     public static boolean isJumping() {
-        return emfEntity instanceof LivingEntity alive && alive.jumping;
+        return IEMFEntity instanceof LivingEntity alive && alive.jumping;
     }
 
 
     public static void setEntityVariable(String variable, float value) {
         //if (variable.equals("var.fly")) System.out.println("setEntityVariable: " + variable + " " + (emfEntity != null ? emfEntity.emf$getVariableMap().getOrDefault(variable, value) : "null"));
-        if (emfEntity != null) {
-            emfEntity.emf$getVariableMap().put(variable, value);
+        if (IEMFEntity != null) {
+            IEMFEntity.emf$getVariableMap().put(variable, value);
         }
     }
 
     public static float getEntityVariable(String variable, float defaultValue) {
         //if (variable.equals("var.fly")) System.out.println("getEntityVariable: " + variable + " " + (emfEntity != null ? emfEntity.emf$getVariableMap().getOrDefault(variable, defaultValue) : "null"));
-        if (emfEntity == null) return defaultValue;
-        return emfEntity.emf$getVariableMap().getOrDefault(variable, defaultValue);
+        if (IEMFEntity == null) return defaultValue;
+        return IEMFEntity.emf$getVariableMap().getOrDefault(variable, defaultValue);
     }
 
     public static void setLayerFactory(Function<ResourceLocation, RenderType> layerFactory) {
@@ -93,8 +114,8 @@ public class EMFAnimationEntityContext {
     }
 
     private static int distanceOfEntityFrom(BlockPos pos) {
-        if (emfEntity == null) return 0;
-        var blockPos = emfEntity.etf$getBlockPos();
+        if (IEMFEntity == null) return 0;
+        var blockPos = IEMFEntity.etf$getBlockPos();
         float f = (float) (blockPos.getX() - pos.getX());
         float g = (float) (blockPos.getY() - pos.getY());
         float h = (float) (blockPos.getZ() - pos.getZ());
@@ -129,7 +150,7 @@ public class EMFAnimationEntityContext {
             lodFactor = factorByFOV;
         }
 
-        if (EMF.config().getConfig().retainDetailOnLargerMobs && emfEntity instanceof Entity entity) {
+        if (EMF.config().getConfig().retainDetailOnLargerMobs && IEMFEntity instanceof Entity entity) {
             var entitySize = Math.max(entity.getBbWidth(), entity.getBbHeight());
             if (entitySize > 2) {
                 lodFactor = (int) (lodFactor / (entitySize / 2));
@@ -149,9 +170,9 @@ public class EMFAnimationEntityContext {
             return true;
         }
 
-        if (EMF.config().getConfig().animationLODDistance == 0 || emfEntity == null) return false;
+        if (EMF.config().getConfig().animationLODDistance == 0 || IEMFEntity == null) return false;
 
-        int lodTimer = lodEntityTimers.getInt(emfEntity.etf$getUuid());
+        int lodTimer = lodEntityTimers.getInt(IEMFEntity.etf$getUuid());
         int lodResult;
         //check lod
         if (lodTimer < 1) {
@@ -159,7 +180,7 @@ public class EMFAnimationEntityContext {
         } else {
             lodResult = lodTimer - 1;
         }
-        lodEntityTimers.put(emfEntity.etf$getUuid(), lodResult);
+        lodEntityTimers.put(IEMFEntity.etf$getUuid(), lodResult);
         lodFrameSkipping = lodResult > 0;
         return lodFrameSkipping;
     }
@@ -184,9 +205,14 @@ public class EMFAnimationEntityContext {
         if (entityIn != null) {
             //perform variant checking for this entity types models
             //this is the only way to keep it generic and also before the entity is rendered and affect al its models
-            Set<Runnable> roots = EMFManager.getInstance().rootPartsPerEntityTypeForVariation.get(entityIn.emf$getTypeString());
+            Set<EMFModelPartRoot> roots = EMFManager.getInstance().rootPartsPerEntityTypeForVariation.get(entityIn.emf$getTypeString());
             if (roots != null) {
-                roots.forEach(Runnable::run);
+                if (isEntityForcedToVanillaModel()) {
+                    roots.forEach((root) -> root.setVariantStateTo(0));
+                } else {
+                    roots.forEach(EMFModelPartRoot::doVariantCheck);
+                }
+
             }
 
             //if this entity requires a debug print do it now after models have variated
@@ -291,7 +317,7 @@ public class EMFAnimationEntityContext {
     }
 
     private static void newEntity(EMFEntity entityIn) {
-        emfEntity = entityIn;
+        IEMFEntity = entityIn;
 
         limbAngle = Float.NaN;
         limbDistance = Float.NaN;
@@ -309,7 +335,7 @@ public class EMFAnimationEntityContext {
     public static void reset() {
         isFirstPersonHand = false;
         layerFactory = null;
-        emfEntity = null;
+        IEMFEntity = null;
         limbAngle = Float.NaN;
         limbDistance = Float.NaN;
         headYaw = Float.NaN;
@@ -344,19 +370,26 @@ public class EMFAnimationEntityContext {
     }
 
     public static float getRuleIndex() {
-        if (emfEntity == null) return 0;
-        return EMFManager.getInstance().lastModelRuleOfEntity.getInt(emfEntity.etf$getUuid());
+        if (IEMFEntity == null) return 0;
+        return EMFManager.getInstance().lastModelRuleOfEntity.getInt(IEMFEntity.etf$getUuid());
+    }
+
+    public static boolean isEntityForcedToVanillaModel(){
+        if (IEMFEntity == null) return false;
+        if (entitiesToForceVanillaModel.contains(IEMFEntity.etf$getUuid())) return true;
+        return EMF.config().getConfig().onlyClientPlayerModel
+                && EMFAnimationEntityContext.getEMFEntity() instanceof Player player && !player.isLocalPlayer();
     }
 
     public static EMFEntity getEMFEntity() {
-        return emfEntity;
+        return IEMFEntity;
     }
 
     public static float getDimension() {
-        if (emfEntity == null || emfEntity.etf$getWorld() == null) {
+        if (IEMFEntity == null || IEMFEntity.etf$getWorld() == null) {
             return 0;
         } else {
-            var optional = emfEntity.etf$getWorld().dimensionTypeRegistration().unwrapKey();
+            var optional = IEMFEntity.etf$getWorld().dimensionTypeRegistration().unwrapKey();
             if (optional.isEmpty()) return 0;
             ResourceLocation id = optional.get().location();
             if (id.equals(BuiltinDimensionTypes.NETHER_EFFECTS)) {
@@ -392,87 +425,87 @@ public class EMFAnimationEntityContext {
     }
 
     public static float getEntityX() {
-        return emfEntity == null ? 0 : (float) Mth.lerp(getTickDelta(), emfEntity.emf$prevX(), emfEntity.emf$getX());
+        return IEMFEntity == null ? 0 : (float) Mth.lerp(getTickDelta(), IEMFEntity.emf$prevX(), IEMFEntity.emf$getX());
     }
 
     public static float getEntityY() {
-        return emfEntity == null ? 0 :
+        return IEMFEntity == null ? 0 :
                 //(double) entity.getY();
-                (float) Mth.lerp(getTickDelta(), emfEntity.emf$prevY(), emfEntity.emf$getY());
+                (float) Mth.lerp(getTickDelta(), IEMFEntity.emf$prevY(), IEMFEntity.emf$getY());
     }
 
     public static float getEntityZ() {
-        return emfEntity == null ? 0 : (float) Mth.lerp(getTickDelta(), emfEntity.emf$prevZ(), emfEntity.emf$getZ());
+        return IEMFEntity == null ? 0 : (float) Mth.lerp(getTickDelta(), IEMFEntity.emf$prevZ(), IEMFEntity.emf$getZ());
     }
 
     public static float getEntityRX() {
-        return (emfEntity == null) ? 0 :
+        return (IEMFEntity == null) ? 0 :
                 //(double) Math.toRadians(entity.getPitch(tickDelta));
-                (float) Math.toRadians(Mth.rotLerp(getTickDelta(), emfEntity.emf$prevPitch(), emfEntity.emf$getPitch()));
+                (float) Math.toRadians(Mth.rotLerp(getTickDelta(), IEMFEntity.emf$prevPitch(), IEMFEntity.emf$getPitch()));
     }
 
     public static float getEntityRY() {
-        if (emfEntity == null) return 0;
-        return (emfEntity instanceof LivingEntity alive) ?
+        if (IEMFEntity == null) return 0;
+        return (IEMFEntity instanceof LivingEntity alive) ?
                 (float) Math.toRadians(Mth.rotLerp(getTickDelta(), alive.yBodyRotO, alive.getVisualRotationYInDegrees())) :
-                emfEntity instanceof Entity entity ?
+                IEMFEntity instanceof Entity entity ?
                         (float) Math.toRadians(Mth.rotLerp(getTickDelta(), entity.yRotO, entity.getYRot()))
                         : 0;
     }
 
     public static float getTime() {
-        if (emfEntity == null || emfEntity.etf$getWorld() == null) {
+        if (IEMFEntity == null || IEMFEntity.etf$getWorld() == null) {
             return 0 + getTickDelta();
         } else {
             //limit value upper limit to preserve floating point precision
-            long upTimeInTicks = emfEntity.etf$getWorld().getGameTime(); // (System.currentTimeMillis() - START_TIME)/50;
+            long upTimeInTicks = IEMFEntity.etf$getWorld().getGameTime(); // (System.currentTimeMillis() - START_TIME)/50;
             return constrainedFloat(upTimeInTicks, 24000) + getTickDelta();
         }
     }
 
     public static float getDayTime() {
-        if (emfEntity == null || emfEntity.etf$getWorld() == null) {
+        if (IEMFEntity == null || IEMFEntity.etf$getWorld() == null) {
             return 0 + getTickDelta();
         } else {
             //limit value upper limit to preserve floating point precision
-            return constrainedFloat(emfEntity.etf$getWorld().getDayTime(), 31415) + getTickDelta();
+            return constrainedFloat(IEMFEntity.etf$getWorld().getDayTime(), 31415) + getTickDelta();
         }
     }
 
     public static float getDayCount() {
-        if (emfEntity == null || emfEntity.etf$getWorld() == null) {
+        if (IEMFEntity == null || IEMFEntity.etf$getWorld() == null) {
             return 0 + getTickDelta();
         } else {
             //limit value upper limit to preserve floating point precision
-            return (float) (emfEntity.etf$getWorld().getDayTime() / 24000L);
+            return (float) (IEMFEntity.etf$getWorld().getDayTime() / 24000L);
         }
     }
 
 
     public static float getHealth() {
-        if (emfEntity == null) return 0;
-        return emfEntity instanceof LivingEntity alive ? alive.getHealth() : 1;
+        if (IEMFEntity == null) return 0;
+        return IEMFEntity instanceof LivingEntity alive ? alive.getHealth() : 1;
     }
 
     public static float getDeathTime() {
-        return emfEntity instanceof LivingEntity alive ? alive.deathTime : 0;
+        return IEMFEntity instanceof LivingEntity alive ? alive.deathTime : 0;
     }
 
     public static float getAngerTime() {
-        if (!(emfEntity instanceof NeutralMob)) return 0;
+        if (!(IEMFEntity instanceof NeutralMob)) return 0;
 
-        float currentKnownHighest = knownHighestAngerTimeByUUID.getInt(emfEntity.etf$getUuid());
-        int angerTime = ((NeutralMob) emfEntity).getRemainingPersistentAngerTime();
+        float currentKnownHighest = knownHighestAngerTimeByUUID.getInt(IEMFEntity.etf$getUuid());
+        int angerTime = ((NeutralMob) IEMFEntity).getRemainingPersistentAngerTime();
 
         //clear anger info if anger is over
         if (angerTime <= 0) {
-            knownHighestAngerTimeByUUID.put(emfEntity.etf$getUuid(), 0);
+            knownHighestAngerTimeByUUID.put(IEMFEntity.etf$getUuid(), 0);
             return 0;
         }
 
         //store this if this is the largest anger time for the entity seen so far
         if (angerTime > currentKnownHighest) {
-            knownHighestAngerTimeByUUID.put(emfEntity.etf$getUuid(), angerTime);
+            knownHighestAngerTimeByUUID.put(IEMFEntity.etf$getUuid(), angerTime);
         }
         return angerTime - getTickDelta();
     }
@@ -480,30 +513,30 @@ public class EMFAnimationEntityContext {
     public static float getAngerTimeStart() {
         //this only makes sense if we are calculating it here from the largest known value of anger time
         // i could also reset it when anger time hits 0
-        return emfEntity instanceof NeutralMob ? knownHighestAngerTimeByUUID.getInt(emfEntity.etf$getUuid()) : 0;
+        return IEMFEntity instanceof NeutralMob ? knownHighestAngerTimeByUUID.getInt(IEMFEntity.etf$getUuid()) : 0;
 
     }
 
     public static float getMaxHealth() {
-        return emfEntity instanceof LivingEntity alive ? alive.getMaxHealth() : 1;
+        return IEMFEntity instanceof LivingEntity alive ? alive.getMaxHealth() : 1;
     }
 
     public static float getId() {
-        return emfEntity == null ? 0 : emfEntity.etf$getUuid().hashCode();
+        return IEMFEntity == null ? 0 : IEMFEntity.etf$getUuid().hashCode();
     }
 
     public static float getHurtTime() {
-        return emfEntity instanceof LivingEntity alive ? (alive.hurtTime > 0 ? alive.hurtTime - getTickDelta() : 0) : 0;
+        return IEMFEntity instanceof LivingEntity alive ? (alive.hurtTime > 0 ? alive.hurtTime - getTickDelta() : 0) : 0;
     }
 
     public static float getHeightAboveGround() {
-        if (!(emfEntity instanceof Entity)) return 0;
+        if (!(IEMFEntity instanceof Entity)) return 0;
         float y = getEntityY();
-        BlockPos pos = emfEntity.etf$getBlockPos();
-        int worldBottom = emfEntity.etf$getWorld().getMinBuildHeight();
+        BlockPos pos = IEMFEntity.etf$getBlockPos();
+        int worldBottom = IEMFEntity.etf$getWorld().getMinBuildHeight();
         //loop down until we hit a block that can be stood on
-        while (!emfEntity.etf$getWorld().getBlockState(pos)
-                .entityCanStandOn(emfEntity.etf$getWorld(),pos, (Entity) emfEntity)
+        while (!IEMFEntity.etf$getWorld().getBlockState(pos)
+                .entityCanStandOn(IEMFEntity.etf$getWorld(),pos, (Entity) IEMFEntity)
                 && pos.getY() > worldBottom) {
             pos = pos.below();
         }
@@ -511,67 +544,67 @@ public class EMFAnimationEntityContext {
     }
 
     public static float getFluidDepthDown() {
-        if (emfEntity == null
-                || emfEntity.etf$getWorld().getFluidState(emfEntity.etf$getBlockPos()).isEmpty()) return 0;
+        if (IEMFEntity == null
+                || IEMFEntity.etf$getWorld().getFluidState(IEMFEntity.etf$getBlockPos()).isEmpty()) return 0;
 
-        BlockPos pos = emfEntity.etf$getBlockPos();
-        int worldBottom = emfEntity.etf$getWorld().getMinBuildHeight();
-        while (!emfEntity.etf$getWorld().getFluidState(pos).isEmpty() && pos.getY() > worldBottom) {
+        BlockPos pos = IEMFEntity.etf$getBlockPos();
+        int worldBottom = IEMFEntity.etf$getWorld().getMinBuildHeight();
+        while (!IEMFEntity.etf$getWorld().getFluidState(pos).isEmpty() && pos.getY() > worldBottom) {
             pos = pos.below();
         }
-        return emfEntity.etf$getBlockPos().getY() - pos.getY();
+        return IEMFEntity.etf$getBlockPos().getY() - pos.getY();
     }
 
     public static float getFluidDepthUp() {
-        if (emfEntity == null
-                || emfEntity.etf$getWorld().getFluidState(emfEntity.etf$getBlockPos()).isEmpty()) return 0;
+        if (IEMFEntity == null
+                || IEMFEntity.etf$getWorld().getFluidState(IEMFEntity.etf$getBlockPos()).isEmpty()) return 0;
 
-        BlockPos pos = emfEntity.etf$getBlockPos();
-        int worldTop = emfEntity.etf$getWorld().getMaxBuildHeight();
-        while (!emfEntity.etf$getWorld().getFluidState(pos).isEmpty() && pos.getY() < worldTop) {
+        BlockPos pos = IEMFEntity.etf$getBlockPos();
+        int worldTop = IEMFEntity.etf$getWorld().getMaxBuildHeight();
+        while (!IEMFEntity.etf$getWorld().getFluidState(pos).isEmpty() && pos.getY() < worldTop) {
             pos = pos.above();
         }
-        return pos.getY() - emfEntity.etf$getBlockPos().getY();
+        return pos.getY() - IEMFEntity.etf$getBlockPos().getY();
     }
 
     public static float getFluidDepth() {
-        if (emfEntity == null
-                || emfEntity.etf$getWorld().getFluidState(emfEntity.etf$getBlockPos()).isEmpty()) return 0;
+        if (IEMFEntity == null
+                || IEMFEntity.etf$getWorld().getFluidState(IEMFEntity.etf$getBlockPos()).isEmpty()) return 0;
         return getFluidDepthDown() + getFluidDepthUp() - 1;
     }
 
     public static boolean isInWater() {
-        return emfEntity != null && emfEntity.emf$isTouchingWater();
+        return IEMFEntity != null && IEMFEntity.emf$isTouchingWater();
     }
 
     public static boolean isBurning() {
-        return emfEntity != null && emfEntity.emf$isOnFire();
+        return IEMFEntity != null && IEMFEntity.emf$isOnFire();
     }
 
     public static boolean isRiding() {
-        return emfEntity != null && emfEntity.emf$hasVehicle();
+        return IEMFEntity != null && IEMFEntity.emf$hasVehicle();
     }
 
     public static boolean isChild() {
-        return emfEntity instanceof LivingEntity alive && alive.isBaby();
+        return IEMFEntity instanceof LivingEntity alive && alive.isBaby();
     }
 
     public static boolean isOnGround() {
-        return emfEntity != null && emfEntity.emf$isOnGround();
+        return IEMFEntity != null && IEMFEntity.emf$isOnGround();
     }
 
     public static boolean isClimbing() {
-        return emfEntity instanceof LivingEntity alive && alive.onClimbable();
+        return IEMFEntity instanceof LivingEntity alive && alive.onClimbable();
     }
 
     public static boolean isAlive() {
-        if (emfEntity == null) return false;
-        return emfEntity.emf$isAlive();
+        if (IEMFEntity == null) return false;
+        return IEMFEntity.emf$isAlive();
     }
 
     public static boolean isUsingHand(boolean right) {
-        if (emfEntity == null) return false;
-        if (emfEntity instanceof LivingEntity entity) {
+        if (IEMFEntity == null) return false;
+        if (IEMFEntity instanceof LivingEntity entity) {
             if(!entity.isUsingItem()) return false;
 
             boolean isRightHanded = entity.getMainArm() == HumanoidArm.RIGHT;
@@ -587,36 +620,36 @@ public class EMFAnimationEntityContext {
     }
 
     public static boolean isAggressive() {
-        if (emfEntity == null) return false;
+        if (IEMFEntity == null) return false;
 
-        if (emfEntity instanceof final EnderMan enderman) {
+        if (IEMFEntity instanceof final EnderMan enderman) {
             return enderman.isCreepy();
         }
-        if (emfEntity instanceof final Blaze blaze) {
+        if (IEMFEntity instanceof final Blaze blaze) {
             return blaze.isOnFire();
         }
-        if (emfEntity instanceof final Guardian guardian) {
+        if (IEMFEntity instanceof final Guardian guardian) {
             return guardian.getActiveAttackTarget() != null;
         }
-        if (emfEntity instanceof final Vindicator vindicator) {
+        if (IEMFEntity instanceof final Vindicator vindicator) {
             return vindicator.isAggressive();
         }
-        if (emfEntity instanceof final SpellcasterIllager caster) {
+        if (IEMFEntity instanceof final SpellcasterIllager caster) {
             return caster.isCastingSpell();
         }
-        if (emfEntity instanceof final NeutralMob angry) {
+        if (IEMFEntity instanceof final NeutralMob angry) {
             return angry.isAngry();
         }
 
-        return emfEntity instanceof Mob mob && mob.isAggressive();
+        return IEMFEntity instanceof Mob mob && mob.isAggressive();
     }
 
     public static boolean isGlowing() {
-        return emfEntity != null && emfEntity.emf$isGlowing();
+        return IEMFEntity != null && IEMFEntity.emf$isGlowing();
     }
 
     public static boolean isHurt() {
-        return emfEntity instanceof LivingEntity alive && alive.hurtTime > 0;
+        return IEMFEntity instanceof LivingEntity alive && alive.hurtTime > 0;
     }
 
     public static boolean isInHand() {
@@ -628,7 +661,7 @@ public class EMFAnimationEntityContext {
     }
 
     public static boolean isInGround() {
-        return is_in_ground_override || emfEntity instanceof Projectile proj && proj.isInWall();
+        return is_in_ground_override || IEMFEntity instanceof Projectile proj && proj.isInWall();
     }
 
     public static boolean isInGui() {
@@ -640,11 +673,11 @@ public class EMFAnimationEntityContext {
     }
 
     public static boolean isInLava() {
-        return emfEntity != null && emfEntity.emf$isInLava();
+        return IEMFEntity != null && IEMFEntity.emf$isInLava();
     }
 
     public static boolean isInvisible() {
-        return emfEntity != null && emfEntity.emf$isInvisible();
+        return IEMFEntity != null && IEMFEntity.emf$isInvisible();
     }
 
     public static boolean isOnHead() {
@@ -660,38 +693,38 @@ public class EMFAnimationEntityContext {
     }
 
     public static boolean isRidden() {
-        return emfEntity != null && emfEntity.emf$hasPassengers();
+        return IEMFEntity != null && IEMFEntity.emf$hasPassengers();
     }
 
     public static boolean isSitting() {
-        if (emfEntity == null) return false;
-        return (emfEntity instanceof TamableAnimal tame && tame.isInSittingPose()) ||
-                (emfEntity instanceof Fox fox && fox.isSitting()) ||
-                (emfEntity instanceof Parrot parrot && parrot.isInSittingPose()) ||
-                (emfEntity instanceof Cat cat && cat.isInSittingPose()) ||
-                (emfEntity instanceof Wolf wolf && wolf.isInSittingPose()) ||
-                (emfEntity instanceof Camel camel && camel.isCamelSitting());
+        if (IEMFEntity == null) return false;
+        return (IEMFEntity instanceof TamableAnimal tame && tame.isInSittingPose()) ||
+                (IEMFEntity instanceof Fox fox && fox.isSitting()) ||
+                (IEMFEntity instanceof Parrot parrot && parrot.isInSittingPose()) ||
+                (IEMFEntity instanceof Cat cat && cat.isInSittingPose()) ||
+                (IEMFEntity instanceof Wolf wolf && wolf.isInSittingPose()) ||
+                (IEMFEntity instanceof Camel camel && camel.isCamelSitting());
     }
 
 
     public static boolean isSneaking() {
-        return emfEntity != null && emfEntity.emf$isSneaking();
+        return IEMFEntity != null && IEMFEntity.emf$isSneaking();
     }
 
     public static boolean isSprinting() {
-        return emfEntity != null && emfEntity.emf$isSprinting();
+        return IEMFEntity != null && IEMFEntity.emf$isSprinting();
     }
 
     public static boolean isTamed() {
-        return emfEntity instanceof TamableAnimal tame && tame.isTame();
+        return IEMFEntity instanceof TamableAnimal tame && tame.isTame();
     }
 
     public static boolean isWet() {
-        return emfEntity != null && emfEntity.emf$isWet();
+        return IEMFEntity != null && IEMFEntity.emf$isWet();
     }
 
     public static float getSwingProgress() {
-        return emfEntity instanceof LivingEntity alive ? alive.getAttackAnim(getTickDelta()) : 0;
+        return IEMFEntity instanceof LivingEntity alive ? alive.getAttackAnim(getTickDelta()) : 0;
     }
 
 //    private static float age = Float.NaN;
@@ -699,10 +732,10 @@ public class EMFAnimationEntityContext {
     public static float getAge() {
 //        if (!Float.isNaN(age)) return age;
 
-        if (emfEntity == null) {
+        if (IEMFEntity == null) {
             return 0 + getTickDelta();
         }else {
-            return constrainedFloat(emfEntity.emf$age(), 24000) + getTickDelta();
+            return constrainedFloat(IEMFEntity.emf$age(), 24000) + getTickDelta();
         }
 //        return age;
     }
@@ -740,7 +773,7 @@ public class EMFAnimationEntityContext {
     }
 
     public static float getLimbAngle() {//limb_swing
-        if (emfEntity == null) return 0;
+        if (IEMFEntity == null) return 0;
         if (Float.isNaN(limbAngle)) {
             doLimbValues();
         }
@@ -752,7 +785,7 @@ public class EMFAnimationEntityContext {
     }
 
     public static float getLimbDistance() {//limb_speed
-        if (emfEntity == null) return 0;
+        if (IEMFEntity == null) return 0;
         if (Float.isNaN(limbDistance)) {
             doLimbValues();
         }
@@ -766,7 +799,7 @@ public class EMFAnimationEntityContext {
     private static void doLimbValues() {
         float o = 0;
         float n = 0;
-        if (!emfEntity.emf$hasVehicle() && emfEntity instanceof LivingEntity alive) {
+        if (!IEMFEntity.emf$hasVehicle() && IEMFEntity instanceof LivingEntity alive) {
             o = alive.walkAnimation.position(getTickDelta());
             n = alive.walkAnimation.speed(getTickDelta());
             if (alive.isBaby()) {
@@ -776,10 +809,10 @@ public class EMFAnimationEntityContext {
             if (n > 1.0F) {
                 n = 1.0F;
             }
-        } else if (emfEntity instanceof Minecart) {
+        } else if (IEMFEntity instanceof Minecart) {
             n = 1;
             o = -(getEntityX() + getEntityZ());
-        } else if (emfEntity instanceof Boat boat) {
+        } else if (IEMFEntity instanceof Boat boat) {
             n = 1;
             //o = boat.interpolatePaddlePhase(0, tickDelta);//1);
             o = Math.max(boat.getRowingTime(1, getTickDelta()), boat.getRowingTime(0, getTickDelta()));
@@ -789,7 +822,7 @@ public class EMFAnimationEntityContext {
     }
 
     public static float getHeadYaw() {
-        if (emfEntity == null) return 0;
+        if (IEMFEntity == null) return 0;
         if (Float.isNaN(headYaw)) {
             doHeadValues();
         }
@@ -801,7 +834,7 @@ public class EMFAnimationEntityContext {
     }
 
     public static float getHeadPitch() {
-        if (emfEntity == null) return 0;
+        if (IEMFEntity == null) return 0;
         if (Float.isNaN(headPitch)) {
             doHeadValues();
         }
@@ -813,7 +846,7 @@ public class EMFAnimationEntityContext {
     }
 
     private static void doHeadValues() {
-        if (emfEntity instanceof LivingEntity livingEntity) {
+        if (IEMFEntity instanceof LivingEntity livingEntity) {
             float h = Mth.rotLerp(getTickDelta(), livingEntity.yBodyRotO, livingEntity.yBodyRot);
             float j = Mth.rotLerp(getTickDelta(), livingEntity.yHeadRotO, livingEntity.yHeadRot);
             float k = j - h;
@@ -867,10 +900,10 @@ public class EMFAnimationEntityContext {
 
 
     public static float getMoveForward() {
-        if (emfEntity == null) return 0;
-        double lookDir = Math.toRadians(90 - emfEntity.emf$getYaw());
+        if (IEMFEntity == null) return 0;
+        double lookDir = Math.toRadians(90 - IEMFEntity.emf$getYaw());
         //float speed = entity.horizontalSpeed;
-        Vec3 velocity = emfEntity.emf$getVelocity();
+        Vec3 velocity = IEMFEntity.emf$getVelocity();
 
         //consider 2d plane of movement with x y
         double x = velocity.x;
@@ -885,10 +918,10 @@ public class EMFAnimationEntityContext {
     }
 
     public static float getMoveStrafe() {
-        if (emfEntity == null) return 0;
-        double lookDir = Math.toRadians(90 - emfEntity.emf$getYaw());
+        if (IEMFEntity == null) return 0;
+        double lookDir = Math.toRadians(90 - IEMFEntity.emf$getYaw());
         //float speed = entity.horizontalSpeed;
-        Vec3 velocity = emfEntity.emf$getVelocity();
+        Vec3 velocity = IEMFEntity.emf$getVelocity();
 
         //consider 2d plane of movement with x y
         double x = velocity.x;