+++ common/src/main/java/traben/entity_model_features/propeties/GlobalVariableFloatProperty.java
@@ -80,8 +80,8 @@ public class GlobalVariableFloatProperty extends RandomProperty {
 
         @Override
         protected @Nullable Float getRangeValueFromEntity(final ETFEntity etfEntity) {
-            if (etfEntity instanceof EMFEntity emfEntity)
-                return getter.apply(emfEntity);
+            if (etfEntity instanceof EMFEntity IEMFEntity)
+                return getter.apply(IEMFEntity);
             return null;
         }