+++ common/src/main/java/traben/entity_model_features/utils/IEMFCuboidDataSupplier.java
@@ -1,6 +1,6 @@
 package traben.entity_model_features.utils;
 
-public interface EMFCuboidDataSupplier {
+public interface IEMFCuboidDataSupplier {
 
 
     int[] emf$getTextureUV();