+++ common/src/main/java/traben/entity_model_features/utils/IEMFWolfCollarHolder.java
@@ -3,7 +3,7 @@ package traben.entity_model_features.utils;
 import net.minecraft.client.model.WolfModel;
 import net.minecraft.world.entity.animal.Wolf;
 
-public interface EMFWolfCollarHolder<T extends Wolf> {
+public interface IEMFWolfCollarHolder<T extends Wolf> {
 
     default boolean emf$hasCollarModel() {
         return emf$getCollarModel() != null;